requirements.txt
streamlit
numpy
huggingface_hub 
pillow
requests
